+++ src/main/java/com/tick42/quicksilver/services/base/RatingService.java
@@ -2,7 +2,7 @@ package com.tick42.quicksilver.services.base;
 
 public interface RatingService {
 
-    int rate(int extensionId, int rating, int userId);
+    double rate(int extensionId, int rating, int userId);
 
     int userRatingForExtension(int extensionId, int userId);
